Support multiline strings in saved config files
When writing out config files such as .configweb, if the values to be saved contain newline characters, mark subsequent lines with leading pipe characters, then reassemble the strings when reading the files back in.

This allows multiline strings to be used for site-wide header and footer values, allowing people to build prettier web interfaces.