Begin every page with a XXSiteHeadXX config marker
Sites might wish to include a standard bit of HTML at the top of every page, site as site-specific navigation links.

This commit adds a XXSiteHeadXX config marker immediately after the <body> tag on every static HTML page and CGI script.

By default, this config string is empty, so applying this commit will not change the appearance of any pages -- it will just add support for this type of site-specific navigation should other users wish to take advantage of it.