Documentation and download link for my.cat
The change updates the "Obtaining the Database" page with links to download my.cat and a new page that documents the category file format.

It also makes a few incidental changes to the "Obtaining the Database" page, which no longer mentions printed versions of the ordinary and armorial, and now accurately reflects the current size of the oanda.db file.

The documentation for my.cat provides a quick outline of the three types of lines it contains, along with some comments on how to interpret each kind of line. There's room for a lot more detail, but this is at least a start.

To enable downloads of the my.cat file, additional questions have been added to the config.web script, with sensible default values.